First cut at new page.tsx organization (left most GUI in old-page.tsx)
